requirements.txt
streamlit>=1.32.0
requests>=2.31.0
tenacity>=8.2.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
pyyaml>=6.0.1
pandas>=2.2.0
openpyxl>=3.1.2
loguru>=0.7.2
python-dotenv>=1.0.1
